New commit

+++ app/Http/Controllers/LawyerController.php
@@ -34,19 +34,18 @@ class LawyerController extends Controller
     public function update(Request $request, $id){
 
         $this->validate($request, ['image' => 'image|mimes:jpeg,png,jpg,gif,svg|max:2048',]);
-        $file = $request->file('image');
-        $file_name = $file->getClientOriginalName();
-
         $lawyer = Lawyer::findOrFail($id);
         $lawyer->gender = $request->gender;
         $lawyer->user->dateOfBirth= $request->dateOfBirth;
         $lawyer->user->city_id = $request->city;
-
-        $upload_folder = '/lawyers/photo/';
-        File::delete(public_path() . $upload_folder. $lawyer->user->photo);
-        $lawyer->user->photo = $file_name;
-        $file->move(public_path() . $upload_folder, $file_name);
-
+        if($request->file('image') != null) {
+            $file = $request->file('image');
+            $file_name = $file->getClientOriginalName();
+            $upload_folder = '/lawyers/photo/';
+            File::delete(public_path() . $upload_folder . $lawyer->user->photo);
+            $lawyer->user->photo = $file_name;
+            $file->move(public_path() . $upload_folder, $file_name);
+        }
         $lawyer->push();
         Session::flash('message', 'Your account updated successfully');
